r/storage_disks_pool: moving into the disks package
