+++ internal/services/disks/storage_disks_pool_resource_test.go
@@ -1,4 +1,4 @@
-package storage_test
+package disks_test
 
 import (
 	"context"
