+++ internal/services/storage/registration.go
@@ -1,7 +1,6 @@
 package storage
 
 import (
-	"github.com/hashicorp/terraform-provider-azurerm/internal/sdk"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
 )
 
@@ -59,15 +58,5 @@ func (r Registration) SupportedResources() map[string]*pluginsdk.Resource {
 		"azurerm_storage_sync":                         resourceStorageSync(),
 		"azurerm_storage_sync_cloud_endpoint":          resourceStorageSyncCloudEndpoint(),
 		"azurerm_storage_sync_group":                   resourceStorageSyncGroup(),
-	}
-}
-
-func (r Registration) DataSources() []sdk.DataSource {
-	return []sdk.DataSource{}
-}
-
-func (r Registration) Resources() []sdk.Resource {
-	return []sdk.Resource{
-		DisksPoolResource{},
 	}
 }
